% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDataToEmeScheme.R
\name{addDataToEmeScheme}
\alias{addDataToEmeScheme}
\title{Title}
\usage{
addDataToEmeScheme(x = emeScheme_gd, s = emeScheme, dataSheet = 1,
  dataCol = 1, verbose = FALSE)
}
\arguments{
\item{x}{path to data file or object containing the data to be filled in. Has to be a
 \code{data.frame} or \code{tibble}. The columns have to be the property
 levels, starting with \code{Property}, and the data column(s). See
 \code{emeScheme_gd} as an example.
 Suppoerted data file formate:
\itemize{
  \item{\code{xlsx} : }{xlsx file in the same format as the file which is loded when excuting \code{enter_new_metadata()} command. It is the best to use this sheet when entering data.}
  \item{others : }{others will be added later}
}}

\item{s}{the \code{emeScheme} object to which the data should be added. If
the data exists, it will be overwritten, \code{NA} and \code{NULL} values
will be treated as not given.}

\item{dataSheet}{if x is an xlsx file, \code{dataSheet} specifies the sheet which
shoiuld be imported. This can be the sheet name or the index. Default is 1.}

\item{dataCol}{data column indicator. Can be an index of data column (1 =
first data column, 2 = second, ...), or the name of the data column.
Default = 1.}

\item{verbose}{if \code{TRUE}, progress and diagnostig messages will be
printed.}
}
\value{
the emeScheme \code{s} filled in with the values as specified in \code{x}
}
\description{
Title
}
\examples{
addDataToEmeScheme( verbose = TRUE )
}
