% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.plantuml.emeSchemeSet.R
\name{as.plantuml.emeSchemeSet}
\alias{as.plantuml.emeSchemeSet}
\title{Generic function to convert \code{emeSchemeSet} to a \code{plantuml} object}
\usage{
as.plantuml.emeSchemeSet(x, complete, nm, ...)
}
\arguments{
\item{x}{object to be converted}

\item{complete}{internal}

\item{nm}{internal}

\item{...}{other arguments passed to other functions}
}
\value{
object of class \code{plantuml} which can be plotted
}
\description{
Generic function to convert \code{emeSchemeSet} to a \code{plantuml} object
}
