% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dmdScheme_raw.R
\name{as_dmdScheme_raw}
\alias{as_dmdScheme_raw}
\title{Generic function to convert the data stored in the object \code{x} into a new object of class \code{dmdScheme_raw...}}
\usage{
as_dmdScheme_raw(x, keepData = FALSE, ..., verbose = FALSE)
}
\arguments{
\item{x}{object to be converted}

\item{keepData}{if the data should be kept or replaced with one row with NAs}

\item{...}{additional arguments for methods}

\item{verbose}{give verbose progress info. Useful for debugging.}
}
\value{
dmdScheme as object of class \code{dmdScheme_raw}
}
\description{
Generic function to convert the data stored in the object \code{x} into a new object of class \code{dmdScheme_raw...}
}
\examples{
as_dmdScheme_raw(dmdScheme, keepData = TRUE)

}
