% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xml.R, R/as_xml.dmdSchemeData.R,
%   R/as_xml.dmdSchemeSet.R
\name{as_xml}
\alias{as_xml}
\alias{as_xml.dmdSchemeData}
\alias{as_xml.dmdSchemeSet}
\title{Generic function to convert an object to xml}
\usage{
as_xml(x, output = "metadata", ...)

\method{as_xml}{dmdSchemeData}(x, output = "metadata", ...)

\method{as_xml}{dmdSchemeSet}(x, output = "metadata", ...)
}
\arguments{
\item{x}{object to be converted.}

\item{output}{specifies the content and format of the exported xml.
\describe{
  \item{"metadata" : }{export of the metadata only with no format attributes}
  \item{"complete" : }{export tof the complete sheme, i.e. "metadata" plus the scheme definition. This is a self contained format which contains all attributes.}
}}

\item{...}{additional arguments for methods}
}
\value{
an \code{xml_document} object
}
\description{
Generic function to convert an object to xml
}
\examples{

x <- as_xml( dmdScheme_example() )
x

## returns \\code{xml_document} object


}
