% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xml_list.R, R/as_xml_list.dmdSchemeSet.R
\name{as_xml_list}
\alias{as_xml_list}
\alias{as_xml_list.dmdSchemeSet}
\title{Generic function to convert an object to a list containing xml(s)}
\usage{
as_xml_list(x, output = "metadata", ...)

\method{as_xml_list}{dmdSchemeSet}(x, output = "metadata", ...)
}
\arguments{
\item{x}{object to be converted.}

\item{output}{specifies the content and format of the exported xml.
\describe{
  \item{"metadata" : }{export of the metadata only with no format attributes}
  \item{"complete" : }{export tof the complete sheme, i.e. "metadata" plus the scheme definition. This is a self contained format which contains all attributes.}
}}

\item{...}{additional arguments for methods}
}
\value{
a \code{list()}  where each element is  \code{xml_document} object
}
\description{
Generic function to convert an object to a list containing xml(s)
}
\examples{

x <- as_xml_list( dmdScheme_example() )
x

## returns a \\code{list()} with one \\code{xml_document} object


}
