% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_ln.R
\name{cat_ln}
\alias{cat_ln}
\title{cat with linefeed at the end Title}
\usage{
cat_ln(...)
}
\arguments{
\item{...}{will be handed over to \code{cat(..., "\n")}}
}
\description{
Copied from tibble:::cat_line()
}
