% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_make_new_scheme.R
\name{dev_make_new_scheme}
\alias{dev_make_new_scheme}
\title{Create anew package skelleton containg the definitions for new domain metadata scheme.}
\usage{
dev_make_new_scheme(schemeDefinition, path = ".")
}
\arguments{
\item{schemeDefinition}{\code{xlsx} Excel file containing the definition of
the metadata scheme. The version and the scheme name should be edited
before the spreadsheet is used here.}

\item{path}{path where the package should be created. Default is the current
working directory.}
}
\value{
invisibly \code{NULL}
}
\description{
\strong{This function is not for the user of a scheme, but for the development process of a new scheme.}
}
\details{
A new metadata scheme can be created as a package which will depend on
\code{dmdScheme}. This function uses the function \code{package.skeleton()} from
the \code{utils} package to create a new directory for the new metadata scheme,
imports the scheme as defined in \code{schemeDefinition} and adds some functions
which make working easier with the new scheme easier. For a documentation of
the workflow to create a new scheme, see the
vignette \strong{Howto Create a new scheme}.
}
\examples{
\dontrun{
dev_make_new_scheme(
  schemeDefinition = "./../emeScheme/emeScheme.xlsx",
  path = tempdir()
)
}
}
