% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.dmdScheme.R
\docType{data}
\name{dmdScheme}
\alias{dmdScheme}
\title{Authorative definition of the dmdScheme.}
\format{An object of class \code{dmdSchemeSet} (inherits from \code{list}) of length 6.}
\usage{
dmdScheme
}
\description{
The dataset contains the authorative definition of the dmdScheme.
There are three S3 classes defined and used:
\describe{
  \item{\code{dmdScheme: }}{all all properties in the scheme have the class
    \code{dmdScheme}} \item{\code{dmdSchemeSet: }}{Property Sets which contain
    other Property Sets have the class \code{dmdSchemeSet}}
  \item{\code{dmdSchemeData: }}{Propertiey Sets which contain only Value
    Properties (in R a \code{tibblle}) have the class \code{dmdSchemeData}}
}
}
\details{
In addition, the last two properties are available with a \code{_NAME} suiffix, specifyinf the name of the Property.
\describe{
  \item{Property Level x}{TODO}
  ...
}
}
\examples{
\dontrun{
## Created by using
dmdScheme <- new_dmdSchemeSet(
  x = dmdScheme_raw,
  keepData = FALSE,
  verbose = TRUE
)
}
}
\keyword{datasets}
