% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmdScheme_installed.R
\name{dmdScheme_installed}
\alias{dmdScheme_installed}
\title{Find all installed dmdSchemes}
\usage{
dmdScheme_installed()
}
\value{
\code{character} matrix in the same format as returned from \code{installed.packages()}
}
\description{
To find all dmdSchemes, this functions looks at all installed packages and
checks if they depend on \code{dmdScheme}. These are identified as
dmdSchemes.
}
\examples{
dmdScheme_installed()

}
