% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.dmdScheme_raw.R
\docType{data}
\name{dmdScheme_raw}
\alias{dmdScheme_raw}
\title{dmdScheme_raw as imported from the Google sheet.}
\format{An object of class \code{dmdSchemeSet_raw} (inherits from \code{list}) of length 6.}
\usage{
dmdScheme_raw
}
\description{
The dataset contains raw data. It was created by using the code below.
}
\examples{
\dontrun{
## Created by using
path <- system.fil;e( "dmdScheme.xlsx", package = "dmdScheme")
dmdScheme_raw <- read_from_excel(
  file = path,
  keepData = FALSE,
  verbose = TRUE,
  raw = TRUE
)
}
}
\keyword{datasets}
