% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmdScheme_to_xml.R
\name{dmdScheme_to_xml}
\alias{dmdScheme_to_xml}
\title{Convert \code{dmdScheme} object to XML}
\usage{
dmdScheme_to_xml(x, tag, file, output = "metadata")
}
\arguments{
\item{x}{\code{dmdScheme} object to be converted}

\item{tag}{xml tag of the root level. \bold{Has to be "dmdScheme for the initial call!}}

\item{file}{empty, \code{NULL} or file name. See details below}

\item{output}{either \code{"metadata"} for export of metadata only or
\code{"complete"} for export including classes et al.}
}
\value{
dependant on the value of \code{file}. See Details
}
\description{
Converts an \code{dmdScheme} object into an xml object.
}
\details{
Depending on the value of \code{file}, the function returns different results:
\itemize{
  \item{\code{file} not given: }{an \code{XMLNode} object}
  \item{\code{file = NULL}: }{a string characterisation of the \code{XMLNode} object}
  \item{\code{file = fileName}: }{saves the string characterisation of the \code{XMLNode} object to an xml file}
}
Based on David LeBauer, Carl Davidson, Rob Kooper. See \url{https://stackoverflow.com/a/27865512/632423}
}
\examples{

## dmdScheme_to_xml( dmdScheme_example, "GoogleData" )
## returns \\code{XMLNode} object

## dmdScheme_to_xml( dmdScheme_example, "GoogleData", file = NULL )
## returns string representation of the \\code{XMLNode} object

}
