% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmdScheme_upgrade.R
\name{dmdScheme_upgrade}
\alias{dmdScheme_upgrade}
\title{Convert older versions to newer versions}
\usage{
dmdScheme_upgrade(x, to = dmdScheme_versions()$scheme)
}
\arguments{
\item{x}{\code{dmdScheme} object}

\item{to}{version to upgrade to. Any version supported is possible, converting to older versions is not supported.}
}
\value{
upgraded \code{dmdScheme} object of version \code{to}
}
\description{
Convert older versions to newer versions
}
