% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmdScheme_versions.R
\name{dmdScheme_versions}
\alias{dmdScheme_versions}
\title{Returns versions of package and loaded dmdScheme}
\usage{
dmdScheme_versions()
}
\value{
a named \code{list}, containing the following objects:
\itemize{
  \item{package}{ containing an object of length on of type \code{package_version} with the version of the package}
  \item{dmdScheme}{ containing an object of length on of type \code{numeric_version} with the version of the metadata scheme dmdScheme}
}
}
\description{
The \code{name} and \code{version} of the \code{scheme}are retrieved from the
loadeed scheme (\code{scheme_use()}), therefore can change when a different
scheme is used. The version of the \code{package} is specific to the package.
}
\examples{
dmdScheme_versions()
dmdScheme_versions(schemeName = "dmdScheme")
}
