% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.emeScheme.R
\docType{data}
\name{emeScheme}
\alias{emeScheme}
\title{Authorative definition of the emeScheme.}
\format{An object of class \code{list} of length 5.}
\usage{
emeScheme
}
\description{
The dataset contains the authorative definition of the emeScheme.
}
\details{
\describe{
  \item{Property Level x}{TODO}
  ...
}
}
\keyword{datasets}
