% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.emeScheme_example.R
\docType{data}
\name{emeScheme_example}
\alias{emeScheme_example}
\title{emeScheme_example as imported from the Google sheet.}
\format{An object of class \code{emeSchemeSet} (inherits from \code{emeScheme}, \code{list}) of length 6.}
\usage{
emeScheme_example
}
\description{
The dataset contains example data. It was created by using the code below.
}
\examples{
\dontrun{
## Created by using
emeScheme_example <- new_emeSchemeSet(
   x = emeScheme_raw,
   keepData = TRUE,
   verbose = TRUE
)
}
}
\keyword{datasets}
