% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeScheme_extract.R
\name{emeScheme_extract}
\alias{emeScheme_extract}
\title{Split \code{emeScheme} object into multiple by using \code{DataFileNameMetaData$dataFileName}}
\usage{
emeScheme_extract(dataFile, x)
}
\arguments{
\item{dataFile}{name of dataFileName whose metadata will be extracted from \code{x}. Has to be an exact match, no wildcards are expanded.}

\item{x}{\code{emeScheme} object from which data to extract}
}
\value{
\code{emeScheme} object containing metadata for the data file \code{dataFileName}
}
\description{
Split \code{emeScheme} object into multiple by using \code{DataFileNameMetaData$dataFileName}
}
