% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.emeScheme_gd.R
\docType{data}
\name{emeScheme_gd}
\alias{emeScheme_gd}
\title{Complete emeScheme Google sheet.}
\format{TODO
\describe{
  \item{Property Level x}{TODO}
  \item{rest}{TODO}
  ...
}}
\source{
\url{https://docs.google.com/spreadsheets/d/1OAyRM1jGL5Vho-YfPffePhwGpQtaBUeTPJV6P98f0Bc}
}
\usage{
emeScheme_gd
}
\description{
The dataset contains the complete imported googlesheet of the emeScheme definition.
}
\keyword{datasets}
