% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeScheme_split.R
\name{emeScheme_split}
\alias{emeScheme_split}
\title{Split \code{emeScheme} object into multiple by using \code{DataFileNameMetaData$dataFileName}}
\usage{
emeScheme_split(x = emeScheme, fileName)
}
\arguments{
\item{x}{\code{emeScheme} object to be split}
}
\value{
\code{list} where each element is one \code{emeScheme} object for a data file as specified in \code{DataFileNameMetaData$dataFileName}
}
\description{
One \code{emeScheme} object can contain metadata for multiple datafiles. For
archiving, these should be split into single \code{emeScheme} objects, one
for each \code{DataFileNameMetaData$dataFileName}.
}
