% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeScheme_upgrade.R
\name{emeScheme_upgrade}
\alias{emeScheme_upgrade}
\title{Convert older versions to newer versions}
\usage{
emeScheme_upgrade(x, to = emeSchemeVersions()$emeScheme)
}
\arguments{
\item{x}{\code{emeScheme} object}

\item{to}{version to upgrade to. Any version supported is possible, converting to older versions is not supported.}
}
\value{
upgraded \code{emeScheme} object of version \code{to}
}
\description{
Convert older versions to newer versions
}
