% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeSchemeToXml.emeSchemeData.R
\name{emeSchemeToXml.emeSchemeData}
\alias{emeSchemeToXml.emeSchemeData}
\title{Convert \code{emeScheme} object to XML}
\usage{
\method{emeSchemeToXml}{emeSchemeData}(x, tag = "emeScheme", file,
  output = "metadata")
}
\arguments{
\item{x}{\code{emeScheme} object to be converted}

\item{tag}{xml tag of the root level. \bold{Has to be "emeScheme for the initial call!}}

\item{file}{empty, \code{NULL} or file name. See details below}

\item{output}{either \code{"metadata"} for export of metadata only or
\code{"complete"} for export including classes et al.}
}
\value{
dependant on the value of \code{file}. See Details
}
\description{
Converts an \code{emeScheme} object into an xml object.
}
\details{
Depending on the value of \code{file}, the function returns different results:
\itemize{
  \item{\code{file} not given: }{an \code{XMLNode} object}
  \item{\code{file = NULL}: }{a string characterisation of the \code{XMLNode} object}
  \item{\code{file = fileName}: }{saves the string characterisation of the \code{XMLNode} object to an xml file}
}
Based on David LeBauer, Carl Davidson, Rob Kooper. See \url{https://stackoverflow.com/a/27865512/632423}
}
\examples{
x <- addDataToEmeScheme()

emeSchemeToXml( x, "GoogleData" )
## returns \\code{XMLNode} object

emeSchemeToXml( x, "GoogleData", file = NULL )
## returns string representation of the \\code{XMLNode} object

}
