% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeSchemeToXml.emeSchemeData_Species.R
\name{emeSchemeToXml.emeSchemeData_Species}
\alias{emeSchemeToXml.emeSchemeData_Species}
\title{Convert \code{emeSchemeData_Species} object to XML}
\usage{
\method{emeSchemeToXml}{emeSchemeData_Species}(x, tag = "emeScheme",
  file, output = "metadata")
}
\description{
Convert \code{emeSchemeData_Species} object to XML
}
