% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeSchemeToXml.emeSchemeData_VideoAnalysis.R
\name{emeSchemeToXml.emeSchemeData_VideoAnalysis}
\alias{emeSchemeToXml.emeSchemeData_VideoAnalysis}
\title{Convert \code{emeSchemeData_VideoAnalysis} object to XML}
\usage{
\method{emeSchemeToXml}{emeSchemeData_VideoAnalysis}(x,
  tag = "emeScheme", file, output = "metadata")
}
\description{
Convert \code{emeSchemeData_VideoAnalysis} object to XML
}
