% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeSchemeToXml.emeSchemeSet.R
\name{emeSchemeToXml.emeSchemeSet}
\alias{emeSchemeToXml.emeSchemeSet}
\title{Convert \code{emeSchemeSet} object to XML}
\usage{
\method{emeSchemeToXml}{emeSchemeSet}(x, tag, file, output = "metadata")
}
\description{
Convert \code{emeSchemeSet} object to XML
}
