% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeSchemeVersions.R
\name{emeSchemeVersions}
\alias{emeSchemeVersions}
\title{Returns versions of package and emeScheme}
\usage{
emeSchemeVersions(lib.loc = NULL)
}
\arguments{
\item{lib.loc}{a character vector of directory names of R libraries, or NULL.
The default value of NULL corresponds to all libraries currently known. If
the default is used, the loaded packages and namespaces are searched before
the libraries.}
}
\value{
a named \code{list()}, containing the following objects:
\itemize{
  \item{package}{ containing an object of length on of type \code{package_version} with the version of the package}
  \item{emeScheme}{ containing an object of length on of type \code{numeric_version} with the version of the metadata scheme emeScheme}
}
}
\description{
Returns versions of package and emeScheme
}
\examples{
emeSchemeVersions()
}
