% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_emeScheme_for_datafile.R
\name{extract_emeScheme_for_datafile}
\alias{extract_emeScheme_for_datafile}
\title{Filter meatdata based on x$DataFile$dataFileName}
\usage{
extract_emeScheme_for_datafile(x, file)
}
\arguments{
\item{x}{object of class \code{emeSchemeSet}}

\item{file}{\code{dataFileName} to be filtered}
}
\value{
object of class \code{emeSchemeSet} with only metadata relevant to the data file as specified in \code{file}.
}
\description{
Extract, based on the \code{emeSchemeSet} DataFile and the names specified there, the data not for the data file will be filtered out.
}
\examples{
extract_emeScheme_for_datafile(emeScheme_example, file = "data1.csv")

}
