% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_emeScheme.R
\name{filter_emeScheme}
\alias{filter_emeScheme}
\title{Internal only to copy classes when using filter}
\usage{
filter_emeScheme(vs, ...)
}
\arguments{
\item{vs}{as iused in \code{dplyr::filter()}}

\item{...}{additional arguments for \code{dplyr::filter()}}
}
\description{
Internal only to copy classes when using filter
}
