% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.microcosmScheme.R
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{microcosmScheme: A package containing the definition for a metadata scheme for =microcosm experiments.}
\description{
The package provides:
1) the authorative definition of the microcosmScheme
2) object definitions for this scheme for R
}
