% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_emeScheme_xlsx.R
\name{format_emeScheme_xlsx}
\alias{format_emeScheme_xlsx}
\title{Format the \code{googlesheet/emeScheme.xlsx}}
\usage{
format_emeScheme_xlsx(fn, emeScheme_gd)
}
\arguments{
\item{fn}{file name where the final xlsx should be saved to. If missing, it will not be saved.}

\item{emeScheme_gd}{specifying this, makes it possible to use this function even during the update process.}
}
\value{
invisibly the workbook as a workbook object as created by \code{xlsx.crerateWorkbook()}
}
\description{
Takes no arguments. Loads  \code{system.file("googlesheet", "emeScheme.xlsx", package = "emeScheme")}, formats it and saves it under the same name.
}
