% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tokens.R
\name{lookup_tokens}
\alias{lookup_tokens}
\title{Replace tokens with vlues from a dmdScheme}
\usage{
lookup_tokens(tokens, scheme, author = "")
}
\arguments{
\item{tokens}{a character vector containing tokens. These can be enclosed in \code{\%\%TOKEN\%\%} or not.}

\item{scheme}{a \code{dmdSchemeSet} object}

\item{author}{the author of the index document}
}
\value{
a list of the length of the input vector \code{tokens}
  containing the objects returned by the tokens. Null if the tokens contains
  invalid values.
}
\description{
For a detailed explanation of these tokens see the vignette \code{Create and
Customize the index Template}.
}
\examples{
lookup_tokens(
  tokens = c(
      "\%\%Treatments.*.2\%\%",
     "\%\%Experiment.*.*\%\%",
     "Measurement.method.3"
   ),
  scheme = dmdScheme_example(),
)
}
