% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example.R
\name{make_example}
\alias{make_example}
\title{Create examples in workingdirectoryt}
\usage{
make_example(name, schemeName = "dmdScheme")
}
\arguments{
\item{name}{name of the example}

\item{schemeName}{name of the package in which the example sits. Default value: \code{dmdScheme}}
}
\value{
invisibly \code{NULL}
}
\description{
This function copies the installed example into the working directory so that it can be used.
}
\examples{
make_example()
\dontrun{
make_example("basic")
}
}
