% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.microcosmScheme.R
\docType{data}
\name{microcosmScheme}
\alias{microcosmScheme}
\title{microcosmScheme as imported from the Google sheet.}
\format{TODO
\describe{
  \item{Property Level x}{TODO}
  ...
}}
\source{
\url{https://docs.google.com/spreadsheets/d/1OAyRM1jGL5Vho-YfPffePhwGpQtaBUeTPJV6P98f0Bc}
}
\usage{
microcosmScheme
}
\description{
The dataset contains the imported and trimmed googlesheet of the microcosmScheme definition.
}
\keyword{datasets}
