% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.microcosmScheme_gd.R
\docType{data}
\name{microcosmScheme_gd}
\alias{microcosmScheme_gd}
\title{Complete microcosmScheme Google sheet.}
\format{TODO
\describe{
  \item{Property Level x}{TODO}
  \item{rest}{TODO}
  ...
}}
\source{
\url{https://docs.google.com/spreadsheets/d/1OAyRM1jGL5Vho-YfPffePhwGpQtaBUeTPJV6P98f0Bc}
}
\usage{
microcosmScheme_gd
}
\description{
The dataset contains the complete imported googlesheet of the microcosmScheme definition.
}
\keyword{datasets}
