% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dmdScheme.R
\name{new_dmdScheme}
\alias{new_dmdScheme}
\title{Generic method to convert the data stored in the object \code{x} into a new object of class \code{dmdScheme...}}
\usage{
new_dmdScheme(x, keepData = FALSE, convertTypes = TRUE,
  verbose = FALSE, warnToError = TRUE, checkVersion = TRUE)
}
\arguments{
\item{x}{object of
\itemize{
\item class \code{dmdSchemeSet_raw} as e.g. returned by \code{read_from_excel(raw = TRUE)}
\item class \code{dmdSchemeData_raw}
}}

\item{keepData}{if the data should be kept or replaced with one row with NAs}

\item{convertTypes}{if \code{TRUE}, the types specified in the types column
are used for the data type. Otherwise, they are left at type
\code{character}}

\item{verbose}{give verbose progress info. Useful for debugging.}

\item{warnToError}{if \code{TRUE}, warnings generated during the conversion
will raise an error}

\item{checkVersion}{if \code{TRUE}, a version mismatch between the package
and the data \code{x} will result in an error. If \code{FALSE}, the check
will be skipped.}
}
\value{
\code{list} of \code{list} of ... \code{tibbles}
}
\description{
Generic method.
}
\examples{
new_dmdScheme(dmdScheme_raw, keepData = TRUE)
new_dmdScheme(dmdScheme_raw$Experiment)

}
