% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_func.R
\name{new_dmdScheme_validation}
\alias{new_dmdScheme_validation}
\title{Create new \code{dmdScheme_validation} object}
\usage{
new_dmdScheme_validation()
}
\value{
new \code{dmdScheme_validation} object
}
\description{
Create new \code{dmdScheme_validation} object
}
\seealso{
Other validation: 
\code{\link{report}()},
\code{\link{valErr_TextErrCol}()},
\code{\link{valErr_extract}()},
\code{\link{valErr_info}()},
\code{\link{valErr_isOK}()},
\code{\link{validate}()}
}
\concept{validation}
