% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dmdSchemeData.R
\name{new_dmdSchemeData}
\alias{new_dmdSchemeData}
\title{Create new \code{dmdSchemeData} class object from specifications}
\usage{
new_dmdSchemeData(x, keepData = TRUE, convertTypes = TRUE,
  verbose = FALSE, warnToError = TRUE)
}
\arguments{
\item{x}{a sheet of the spreadsheet displayed by calling
\code{enter_new_metadata()} converted to a \code{data.frame} or
\code{tibble} by calling \code{readxl::read_excel()}. Class has to be \code{dmdSchemeData_raw}.}

\item{keepData}{if the data should be trimmed to one row with NAs}

\item{convertTypes}{if \code{TRUE}, the types specified in the types column
are used for the data type. Otherwise, they are left at type \code{character}}

\item{verbose}{give verbose progress info. Useful for debugging.}

\item{warnToError}{if \code{TRUE}, warnings generated during the conversion will raise an error}
}
\value{
\code{dmdSchemeData} Data object
}
\description{
Usually only called from within the function \code{new_dmdSchemeSet()}.
}
\examples{
new_dmdSchemeData(dmdScheme_raw$Experiment)

}
