% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_emeSchemeData.R
\name{new_emeSchemeData}
\alias{new_emeSchemeData}
\title{Create new \code{emeSehemeData} class object from specifications}
\usage{
new_emeSchemeData(x, trimToOne = TRUE, debug = FALSE)
}
\arguments{
\item{x}{tibble as imported from the google sheet}
}
\value{
\code{emeSchemeData} object complete with all attributes
}
\description{
Create new \code{emeSehemeData} class object from specifications
}
\examples{
new_emeSchemeData( x = emeSCheme_gd)
}
