% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_emeSchemeSet.R
\name{new_emeSchemeSet}
\alias{new_emeSchemeSet}
\title{Convert the data stored in \code{emeScheme_raw} into a list of tibbles}
\usage{
new_emeSchemeSet(x, keepData = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{object of class \code{emeSchemeSet_raw} as e.g. returned by \code{read_from_excel(raw = TRUE)}}

\item{keepData}{if the data should be kept or replaced with one row with NAs}

\item{verbose}{give messages to make finding errors in data easier}
}
\value{
\code{list} of \code{list} of ... \code{tibbles}
}
\description{
Convert the data stored in \code{emeScheme_raw} into a list of tibbles
}
\examples{
new_emeSchemeSet(emeScheme_raw, keepData = TRUE)

}
