% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.emeScheme.R
\docType{package}
\name{pkg.emeScheme}
\alias{pkg.emeScheme}
\alias{pkg.emeScheme-package}
\title{emeScheme: A package containing the definition for a metadata scheme for Experimental Microcosm Ecological experiments.}
\description{
The package provides:
1) the authorative definition of the emeScheme
2) object definitions for this scheme for R
}
