% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dmdSchemeData.R
\name{print.dmdSchemeData}
\alias{print.dmdSchemeData}
\title{Print method for \code{dmdSchemeData} object}
\usage{
\method{print}{dmdSchemeData}(
  x,
  ...,
  printAttr = TRUE,
  printExtAttr = FALSE,
  printData = TRUE,
  .prefix = ""
)
}
\arguments{
\item{x}{object of type \code{dmdSchemeSet}}

\item{...}{additional arguments - not used here}

\item{printAttr}{default \code{TRUE} - attributes are printed}

\item{printExtAttr}{default \code{FALSE} - additional attributes are not printed (e.g. \code{class})}

\item{printData}{default \code{TRUE} - data is printed}

\item{.prefix}{mainly for internal use - prefix for all printed lines}
}
\value{
invisibly x
}
\description{
Print method for \code{dmdSchemeData} object
}
