% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dmdSchemeSet.R
\name{print.dmdSchemeSet}
\alias{print.dmdSchemeSet}
\title{Print method for \code{dmdSchemeSet} object}
\usage{
\method{print}{dmdSchemeSet}(x, ..., printAttr = TRUE,
  printExtAttr = FALSE, printData = TRUE, .prefix = "")
}
\arguments{
\item{x}{object of type \code{dmdSchemeSet}}

\item{...}{additional arguments - not used here}

\item{printAttr}{default \code{TRUE} - attributes are printed}

\item{printExtAttr}{default \code{FALSE} - additional attributes are not printed (e.g. \code{class})}

\item{printData}{default \code{TRUE} - data is printed}

\item{.prefix}{mainly for internal use - prefix for all printed lines}
}
\value{
invisibly x
}
\description{
Print method for \code{dmdSchemeSet} object
}
