% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.emeSchemeSet.R
\name{print.emeSchemeSet}
\alias{print.emeSchemeSet}
\title{Generic print function for \code{emeSchemeSet}}
\usage{
\method{print}{emeSchemeSet}(x, printAttr = TRUE, printExtAttr = FALSE,
  printData = TRUE, .prefix = "")
}
\arguments{
\item{x}{object of type \code{emeSchemeSet}}

\item{printAttr}{default \code{TRUE} - attributes are printed}

\item{printExtAttr}{default \code{FALSE} - additional attributes are not printed (e.g. \code{class})}

\item{printData}{default \code{TRUE} - data is printed}

\item{.prefix}{mainly for internal use - prefix for all printed lines}
}
\value{
invisibly x
}
\description{
Generic print function for \code{emeSchemeSet}
}
