% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dmdScheme_validation.R
\name{print_dmdScheme_validation_summary}
\alias{print_dmdScheme_validation_summary}
\title{Internal function to ptint dmdScheme_validation of format \code{summary}}
\usage{
print_dmdScheme_validation_summary(
  x,
  level,
  listLevel,
  error = c(0, 1, 2, 3, NA)
)
}
\arguments{
\item{x}{as in print.dmdScheme_validation}

\item{level}{as in print.dmdScheme_validation}

\item{listLevel}{as in print.dmdScheme_validation}

\item{error}{numeric v ector, containing error levels to print. Default is all error levels.}
}
\value{
as in print.dmdScheme_validation
}
\description{
Internal function to ptint dmdScheme_validation of format \code{summary}
}
