% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.emeScheme_validation.R
\name{print_emeScheme_validation_summary}
\alias{print_emeScheme_validation_summary}
\title{Internal function to ptint emeScheme_validation of format \code{summary}}
\usage{
print_emeScheme_validation_summary(x, level, listLevel)
}
\arguments{
\item{x}{as in print.emeScheme_validation}

\item{level}{as in print.emeScheme_validation}

\item{listLevel}{as in print.emeScheme_validation}
}
\value{
as in print.emeScheme_validation
}
\description{
Internal function to ptint emeScheme_validation of format \code{summary}
}
