% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.propSets.R
\docType{data}
\name{propSets}
\alias{propSets}
\title{The sheets in the excel workbook, i.e. the dmdSchemeData property sets}
\format{An object of class \code{character} of length 6.}
\usage{
propSets
}
\description{
The sheets in the excel workbook, i.e. the dmdSchemeData property sets
}
\examples{
propSets

}
\keyword{datasets}
