% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_raw.R
\name{read_excel_raw}
\alias{read_excel_raw}
\title{Read scheme data from Excel file into \code{\link[dmdScheme]{dmdScheme_raw}} object}
\usage{
read_excel_raw(file, verbose = FALSE, checkVersion = TRUE)
}
\arguments{
\item{file}{the name of the Excel file (.xls or .xlsx) containing the data to be
read.}

\item{verbose}{give verbose progress info. Useful for debugging.}

\item{checkVersion}{if \code{TRUE}, check for version or scheme conflicts
between the package scheme version and the scheme version of the file.
Aborts with an error if there is a conflict.}
}
\value{
object of class \code{dmdScheme_raw}
}
\description{
Reads the data from an Excel file as is and no validation. Only validation of
the scheme version and scheme name is done (when \code{checkVersion = TRUE}).
}
\examples{
read_excel_raw( scheme_path_xlsx() )

}
