% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_excel.R
\name{read_from_excel}
\alias{read_from_excel}
\title{Read scheme data from Excel file}
\usage{
read_from_excel(file, keepData = TRUE, verbose = FALSE, raw = FALSE)
}
\arguments{
\item{file}{the name of the Excel file (.xls or .xlsx) containing the data to be
read from.}

\item{keepData}{if the data in \code{file} should be kept or replaced with
one row with NAs. \code{keepData = FALSE} is only importing the structure
of the \code{emeScheme} as in the variable \code{emeScheme}.}

\item{verbose}{give verbose progress info. Useful for debugging.}

\item{raw}{if \code{TRUE} the excel file will be read as-is and not converted
to an \code{emeScheme} object.}
}
\value{
either if \code{raw = TRUE} a list of tibbles from the worksheets as
  defined in \code{propSets} of Class \code{emeScheme_raw}, otgherwise an
  object of class \code{emeSchemeSet}
}
\description{
Reads the data from an Excel file. TRhe structure of the Excel file has to be
\bold{identical} to the one opened by \code{enter_new_metadata()}
}
\examples{
read_from_excel(file = system.file("googlesheet", "emeScheme.xlsx", package = "emeScheme"))

}
