% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_excel_raw.R
\name{read_from_excel_raw}
\alias{read_from_excel_raw}
\title{Read scheme data from Excel file}
\usage{
read_from_excel_raw(file, keepData = TRUE, verbose = FALSE,
  schemeName = "dmdScheme", checkVersion = TRUE)
}
\arguments{
\item{file}{the name of the Excel file (.xls or .xlsx) containing the data to be
read from.}

\item{keepData}{if the data in \code{file} should be kept or replaced with
one row with NAs. \code{keepData = FALSE} is only importing the structure
of the \code{dmdScheme} as in the variable \code{dmdScheme}.}

\item{verbose}{give verbose progress info. Useful for debugging.}

\item{schemeName}{name of the scheme. Default: dmdScheme. Only for developing
new schemes needed.}

\item{checkVersion}{if \code{TRUE}, abort if version conflicts between the
package scheme version and the scheme version of the file.}
}
\value{
either if \code{raw = TRUE} a list of tibbles from the worksheets of
  Class \code{dmdScheme_raw}, otherwise an object of class
  \code{dmdSchemeSet}
}
\description{
Reads the data from an Excel file. TRhe structure of the Excel file has to be
\bold{identical} to the one opened by \code{enter_new_metadata()}
}
\examples{
read_from_excel(file = system.file("dmdScheme.xlsx", package = "dmdScheme"))

}
