% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports_eml.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{write_eml}
\alias{eml_validate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{EML}{\code{\link[EML]{eml_validate}}, \code{\link[EML]{write_eml}}}
}}

