% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run shiny app.}
\usage{
run_app()
}
\value{
return value from \code{runApp()}
}
\description{
The shiny app allows the entering, validating, and exporting of the metadata
without using R.
See https://deanattali.com/2015/04/21/r-package-shiny-app/
}
\examples{
\dontrun{
run_app()
}

}
