% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme_install.R, R/scheme_list.R,
%   R/scheme_use.R
\name{scheme_install}
\alias{scheme_install}
\alias{scheme_list}
\alias{scheme_use}
\title{Add \code{schemeDefinition} file to installed schemes}
\usage{
scheme_install(schemeDefinition, overwrite = FALSE)

scheme_list()

scheme_use(schemeDefinition = "dmdScheme.xml")
}
\arguments{
\item{schemeDefinition}{file containing the scheme definition in a recognized format}

\item{overwrite}{if \code{TRUE}, the scheme will be overwritten if it exists}

\item{schemeDefinition}{}
}
\value{
result of \link{file.copy()}

list of schemes installed in the folder \code{system.file("schemeDefinitions", package = "dmdScheme")}
}
\description{
Installed schemes are copied to \code{system.file("schemeDefinitions",
package = "dmdScheme")} and , if necessary, an \code{.xlsx} definition is
saved in addition. These can be listed by using \link{scheme_list()}.

Schemes can be installed by using the \link{scheme_install()} function.

Switch from the current scheme to a new scheme as defined in the file
\code{schemeDefinition}. The name of the scheme definition as in the folder \code{}.
New schemes can be added to the library via a call to \code{scheme_add()}.
}
\examples{
scheme_list()
use_scheme("dmdScheme.xml)
use_scheme("dmdScheme.xlsx)
\dontrun{
scheme_install("path/to/definition.xml)
scheme_install("path/to/definition.xlsx)
}
}
