% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.scheme_active.R
\docType{data}
\name{scheme_active}
\alias{scheme_active}
\title{Name and version of the active scheme.}
\format{An object of class \code{NULL} of length 0.}
\usage{
scheme_active
}
\description{
Name and version of the active scheme.
}
\keyword{datasets}
