% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme_add.R
\name{scheme_add}
\alias{scheme_add}
\title{Add \code{schemeDefinition} file to installed schemes}
\usage{
scheme_add(schemeDefinition, overwrite = FALSE)
}
\arguments{
\item{schemeDefinition}{file containing the scheme definition in a recognized format}

\item{overwrite}{if \code{TRUE}, the scheme will be overwritten if it exists}
}
\value{
result of \link{file.copy()}

#' @rdname scheme
}
\description{
Installed schemes are copied to \code{system.file("schemeDefinitions",
package = "dmdScheme")} and , if necessary, an \code{.xlsx} definition is
saved in addition. These can be listed by using \link{scheme_list()}.
}
