% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme_make.R
\name{scheme_make}
\alias{scheme_make}
\title{Functions to manage schemes}
\usage{
scheme_make(schemeDefinition, path = ".", overwrite = FALSE)
}
\arguments{
\item{schemeDefinition}{path to the \code{.xlsx} file containing the \bold{definition} of the scheme as well as the \bold{example}}

\item{path}{where the final scheme definition should be created.}

\item{overwrite}{if \code{TRUE}, the scheme definition in \code{path} will be overwritten.}
}
\value{
fully qualified path to the created scheme
}
\description{
Functions to manage schemes
}
