% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme_path_index_template.R
\name{scheme_path_index_template}
\alias{scheme_path_index_template}
\title{Functions to manage schemes}
\usage{
scheme_path_index_template()
}
\value{
fully qualified path to the \code{index} template file
}
\description{
Functions to manage schemes
}
\examples{
scheme_path_index_template()

}
