% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme_path_xml.R
\name{scheme_path_xml}
\alias{scheme_path_xml}
\title{Functions to manage schemes}
\usage{
scheme_path_xml()
}
\value{
fully qualified path to the \code{xml} file containing the scheme definitipn and the example, \code{NULL} if it does not exist.
}
\description{
Functions to manage schemes
}
\examples{
scheme_path_xlsx()

}
