% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_emeScheme.R
\name{select_emeScheme}
\alias{select_emeScheme}
\title{Internal only to copy classes when using select}
\usage{
select_emeScheme(vs, ...)
}
\arguments{
\item{vs}{as iused in \code{dplyr::filter()}}

\item{...}{additional arguments for \code{dplyr::filter()}}
}
\description{
Internal only to copy classes when using select
}
