% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet_upgrade.R
\name{spreadsheet_upgrade}
\alias{spreadsheet_upgrade}
\title{Convert older versions to newer versions}
\usage{
spreadsheet_upgrade(x, to = dmdSchemeVersions()$scheme)
}
\arguments{
\item{x}{file name of spreadsheet}

\item{to}{version to upgrade to. Any version supported is possible, converting to older versions is not supported.}
}
\value{
file name of upgraded spreadsheet (\code{x_to.xlsx} where \code{x} is the original file name and \code{to} is the new version)
}
\description{
Convert older versions to newer versions
}
