% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTranspose.R
\name{toTranspose}
\alias{toTranspose}
\title{Return tabs in scheme definition in Excel document which need to be
transposed or if a tab has to be transposed}
\usage{
toTranspose(tabs = NULL)
}
\arguments{
\item{tabs}{if \code{NULL}, the}
}
\value{
if \code{tabs} is null, the \code{character} vector containing the
  sheets which need to be transposed. If \code{tabs} is not NULL, a
  \code{logical} vector of the same length as \code{tabs} which is
  \code{TRUE}, if the tab needs to be transposed, otherwise \code{FALSE}.
}
\description{
Tabs in the Excel document need to be, for processing, transposed, if they
are vertical as the Experiment tab is.
}
\examples{
toTranspose()
# [1] "Experiment"     "MdBibliometric"

toTranspose(c("Experiment", "Not"))
# [1]  TRUE FALSE

}
