% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_from_new_sheet.R
\name{update_from_new_sheet}
\alias{update_from_new_sheet}
\title{Update data from dmdScheme.xlsx --- ONLY FOR DEVLOPMENT NEEDED}
\usage{
update_from_new_sheet(newDmdScheme, updateSchemeVersion = TRUE,
  updatePackageName = FALSE)
}
\arguments{
\item{newDmdScheme}{xlsx spreadsheet containing the new \code{dmdScheme}
definition has the same md5 suma s in the DESCRIPTION}

\item{updateSchemeVersion}{if \code{TRUE}, the field \code{dmdSchemeVersion}
in the DESCRIPOTION file is updated. Default: \code{TRUE}}

\item{updatePackageName}{if \code{TRUE}, the field \code{Package} in the
DESCRIPOTION file is updated. This should be used with caution, as it might
need other changes in the poackage. Only useful for the first import.
Default: \code{FALSE}}
}
\value{
invisibly NULL
}
\description{
Update the data from the file \code{file.path( ".", "inst",
"dmdScheme.xlsx")} and bump the version in the DESCRIPTION if it has changed.
Creates versioned copy
Backup - should be deleted when packaging.
}
