% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_old_files.R
\name{upgrade_old_files}
\alias{upgrade_old_files}
\title{Convert older versions of to newer versions}
\usage{
upgrade_old_files(file, to = dmdScheme_versions()$scheme)
}
\arguments{
\item{file}{file name of spreadsheet or of exported xml file}

\item{to}{version to upgrade to. Any version supported is possible, converting to older versions is not supported.}
}
\value{
file name of upgraded spreadsheet (\code{BASENAME(x).to.EXTENSION(x)} where \code{x} is the original file name and \code{to} is the new version)
}
\description{
The new file will have new version as a suffix.
}
\examples{
\dontrun{
upgrade("dmdScheme.xlsx")
upgrade("dmdScheme.xml")
}
}
