% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationErrorFunctions.R
\name{valErr_extract}
\alias{valErr_extract}
\title{Extract all fields named object of class \code{emeScheme_validation}}
\usage{
valErr_extract(x)
}
\arguments{
\item{x}{object of class \code{emeScheme_validation}}
}
\value{
named numeric vector of the error levels of the different validations done
}
\description{
Extract all fields named object of class \code{emeScheme_validation}
}
