% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationErrorFunctions.R
\name{valErr_info}
\alias{valErr_info}
\title{Return info about error representation}
\usage{
valErr_info(error)
}
\arguments{
\item{error}{either level, text or colour of error (see \code{valErr_errorLevels})}
}
\value{
the row from valErr_errorLevels corresponding to the argument \code{error}
}
\description{
Return info about error representation
}
