% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_error_functions.R
\name{valErr_isOK}
\alias{valErr_isOK}
\title{Creates \code{data.frame} from object of class \code{dmdScheme_validation} for usage in \code{details} of validation}
\usage{
valErr_isOK(x, returnRootError = FALSE)
}
\arguments{
\item{x}{\code{data.frame} with the fields \code{Module}, \code{error} and \code{isOK}}

\item{returnRootError}{if \code{TRUE}, return all errors \bold{including} the error in the object x.}
}
\value{
named numeric vector of the error levels of the different validations done
}
\description{
Creates \code{data.frame} from object of class \code{dmdScheme_validation} for usage in \code{details} of validation
}
