% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationErrorFunctions.R
\name{valErr_TextErrCol}
\alias{valErr_TextErrCol}
\title{Colour the \code{text} by using the error colour}
\usage{
valErr_TextErrCol(text, error, addError = TRUE)
}
\arguments{
\item{text}{to be coloured. if not supplied, the coloured error text will be
returned. If \code{text} is of class \code{emeScheme_validation}, the
function will be called with \code{text = text$header, error = text$error}}

\item{error}{either level, text or colour of error (see \code{valErr_errorLevels})}

\item{addError}{if the error text should be added in the front of the \code{text}.}
}
\value{
the coloured text or error text
}
\description{
Colour the \code{text} by using the error colour
}
