% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate structure of object of class \code{emeSchemeSet_raw} against \code{emeScheme}.}
\usage{
validate(x, path = "", validateData = TRUE, errorIfFalse = TRUE)
}
\arguments{
\item{x}{object of class \code{emeSchemeSet_raw} as returned from \code{read_from_excel( keepData = FALSE, raw = TRUE)}}

\item{path}{path to the data files}

\item{validateData}{if \code{TRUE} data is validated as well; the structure is always validated}

\item{errorIfFalse}{if \code{TRUE} an error will be raised if the schemes are not identical, i.e. there are structural differences.}
}
\value{
\code{TRUE} if they are identical or character vector describing the differences
}
\description{
Validate structure of object of class \code{emeSchemeSet_raw} against \code{emeScheme}.
}
\examples{
validate( emeScheme_raw )

}
