% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_raw.R
\name{validate_raw}
\alias{validate_raw}
\title{Validate structure of object of class \code{emeSchemeSet_raw} against \code{emeScheme}.}
\usage{
validate_raw(x, errorIfFalse = TRUE)
}
\arguments{
\item{x}{object of class \code{emeSchemeSet_raw} as returned from \code{read_from_excel( keepData = FALSE, raw = TRUE)}}

\item{errorIfFalse}{if \code{TRUE} an error will be raised if the schemes are not identical, i.e. there are structural differences.}
}
\value{
\code{TRUE} if they are identical or character vector describing the differences
}
\description{
Validate structure of object of class \code{emeSchemeSet_raw} against \code{emeScheme}.
}
\examples{
validate_raw( emeScheme_raw )

}
