% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_report.R
\name{validate_report}
\alias{validate_report}
\title{Validates \code{x} and creates report}
\usage{
validate_report(x, path = ".", file, open = TRUE, report = "html",
  report_author = "Tester",
  report_title = "Validation of data against dmdScheme")
}
\arguments{
\item{x}{object of class \code{dmdSchemeSet_raw} as returned from
\code{read_from_excel( keepData = FALSE, raw = TRUE)} or file name of an
xlsx file containing the metadata.}

\item{path}{path to the data files}

\item{file}{name of the file containing the generated report, including
extension. If missing, ot will be saved as a temporary file in the
temporary folder.}

\item{open}{if \code{TRUE}, open the report. Default: \code{TRUE}}

\item{report}{determines if and in which format a report of the validation
  should be generated. Allowed values are:
\itemize{
  \item{\bold{'none'}   : } {no report is generated}
  \item{\bold{'html'}   : } {a html (.html) report is generated and opened}
  \item{\bold{'pdf'}    : } {a pdf (.pdf) report is generated and opened}
  \item{\bold{'word'}   : } {a word (.docx)report is generated and opened}
}}

\item{report_author}{name of the author for the report}

\item{report_title}{title of the report for the report}
}
\value{
return the \code{dmdScheme_validation} object
}
\description{
This function validates either an excel workbook or an object of class
\code{dmdScheme_raw} against the definition of the \code{dmdScheme}. It
validates the structure, the types, suggested values, ... and returns an object of class \code{dmdScheme_validation}.
TODO: DESCRIBE THE CLASS!
Based on this result, a report is created if asked fore (see below for details).
}
\examples{
validate_report( dmdScheme_raw )
\dontrun{
validate_report(
   x = dmdScheme_raw,
   report = "html",
   report_author = "The Author I am",
   report_title = "A Nice Report"
)
}

}
