% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_func.R
\name{validateAllowedValues}
\alias{validateAllowedValues}
\title{Validate allowed values}
\usage{
validateAllowedValues(sraw)
}
\arguments{
\item{sraw}{object of type \code{dmdScheme_data} generated with types not converted}
}
\value{
\code{dmdScheme_validation} object
}
\description{
Validate allowed values
}
