% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_func.R
\name{validateDataFileMetaDataDataFileExists}
\alias{validateDataFileMetaDataDataFileExists}
\title{Validate suggested values}
\usage{
validateDataFileMetaDataDataFileExists(xraw, path)
}
\arguments{
\item{xraw}{object of type \code{dmdScheme_set} generated with types not converted}

\item{path}{path to the data files}
}
\value{
\code{dmdScheme_validation} object
}
\description{
Validate suggested values
}
