% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_func.R
\name{validateStructure}
\alias{validateStructure}
\title{Validate structure of \code{dmdScheme} object}
\usage{
validateStructure(x)
}
\arguments{
\item{x}{object of type \code{dmdScheme_raw}}
}
\value{
\code{dmdScheme_validation} object
}
\description{
Validate structure of \code{dmdScheme} object
}
