% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_func.R
\name{validateSuggestedValues}
\alias{validateSuggestedValues}
\title{Validate suggested values}
\usage{
validateSuggestedValues(sraw)
}
\arguments{
\item{sraw}{object of type \code{dmdScheme_data} generated with types not converted}
}
\value{
\code{dmdScheme_validation} object
}
\description{
Validate suggested values
}
