% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_func.R
\name{validateTypes}
\alias{validateTypes}
\title{Validate type of tab}
\usage{
validateTypes(sraw, sconv)
}
\arguments{
\item{sraw}{object of type \code{dmdScheme_data} generated with types not converted}

\item{sconv}{object of type \code{dmdScheme_data} generated with types converted}
}
\value{
\code{dmdScheme_validation} object
}
\description{
Validate type of tab
}
