% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel.R
\name{write_excel}
\alias{write_excel}
\title{Write write x as an excel file to disk}
\usage{
write_excel(x, file, ...)
}
\arguments{
\item{x}{object which can be convertewd to a \code{dmdSchemeSet_raw} odject using the function \code{as_dmdScheme_raw} which will
saved as an \code{xlsx} file.}

\item{file}{Path to file or connection to write to.}

\item{...}{additional parameter for the conversion function \code{(openxlsx::write.xlsx())}}
}
\value{
invisibly returns the path to the file saved to
}
\description{
Convert the object \code{x} to an \code{xml_document} object using the function \code{as_xml()} and write it to a file.
If no method \code{as_xml()} exists for the object |class{x}, an error will be raised.
}
\examples{
write_excel(dmdScheme(), file = tempfile())
write_excel(dmdScheme_raw(), file = tempfile())

}
