% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml.R
\name{write_xml}
\alias{write_xml}
\title{Function to write x as an XML file to disk}
\usage{
write_xml(x, file)
}
\arguments{
\item{x}{object which will be coinverted and saved as an xml file.}

\item{file}{Path to file or connection to write to.}

\item{...}{additional parameter for the conversion function \code{as_xml()}}
}
\description{
Convert the object \code{x} to an \code{xml_document} object using the function \code{as_xml()} to a file.
}
\examples{
write_xml(dmdScheme_raw, file = tempfile())

}
