% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_dmdScheme.R
\name{xml_to_dmdScheme}
\alias{xml_to_dmdScheme}
\title{Convert a exported xml file to a \code{dmdScheme_Set} object}
\usage{
xml_to_dmdScheme(x, onlyScheme = FALSE, useSchemeInXml = NULL,
  verbose = FALSE)
}
\arguments{
\item{x}{a file containing the xml or an \code{xml_document} object}

\item{useSchemeInXml}{if \code{TRUE} use scheme definition in xml and raise
an error if the xml does not contain a scheme definition. If False, use the scheme definition from the corresponding
installed package, even if the xml contains a scheme definition. if
\code{NULL} (the default), use the definition in the xml if it is contains a definition,
if not use the corresponding definition from the installed package.}

\item{verbose}{give verbose progress info. Useful for debugging.}
}
\value{
\code{dmdScheme} or descendant object
}
\description{
The scheme is automatically chosen from the propety \code{propertyName} in
the xml file. Nevertheless, the scheme has to be installed, but it does not
hav to be loaded. An error is raised if it does not exist.
}
\examples{
## x <- dmdScheme_example

}
