% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_dmdSchemeOnly.R
\name{xml_to_dmdSchemeOnly}
\alias{xml_to_dmdSchemeOnly}
\title{Convert a exported xml file to a \code{dmdScheme} scheme, ignoring the included data}
\usage{
xml_to_dmdSchemeOnly(x)
}
\arguments{
\item{x}{a file containing the xml (exported with \code{output = "complete"}
or an \code{xml_document} object (from the \code{xml2} package)}
}
\value{
\code{dmdScheme} or descendant object
}
\description{
Convert a exported xml file to a \code{dmdScheme} scheme, ignoring the included data
}
