% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_upgrade.R
\name{xml_upgrade}
\alias{xml_upgrade}
\title{Convert older versions to newer versions}
\usage{
xml_upgrade(x, to = dmdScheme_versions()$scheme)
}
\arguments{
\item{x}{file name of xml file}

\item{to}{version to upgrade to. Any version supported is possible, converting to older versions is not supported.}
}
\value{
file name of upgraded xml file (\code{x_to.xml} where \code{x} is the original file name and \code{to} is the new version)
}
\description{
Convert older versions to newer versions
}
