% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlToEmeScheme.R
\name{xmlToEmeScheme}
\alias{xmlToEmeScheme}
\title{Convert xml to \code{emeScheme}}
\usage{
xmlToEmeScheme(xml)
}
\arguments{
\item{xml}{either xml object or text reresenatation from text object}
}
\value{
\code{emeScheme} object
}
\description{
The resulting \code{emeScheme} does only contain the non-missing values which
are specified in the xml file.
}
\examples{
x <- addDataToEmeScheme()

y <- emeSchemeToXml( x, "GoogleData" )
xmlToEmeScheme( y )
## works with an \\code{XMLNode} object

y <- emeSchemeToXml( x, "GoogleData", file = NULL )
xmlToEmeScheme( y )
## works with the string representation of an \\code{XMLNode} object

}
