% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDataToEmeScheme.R
\name{addDataToEmeScheme}
\alias{addDataToEmeScheme}
\title{Title}
\usage{
addDataToEmeScheme(x = emeScheme_gd, s = emeScheme, dataCol = 1,
  verbose = FALSE)
}
\arguments{
\item{x}{object containing the data to be filled in. Has to be a
\code{data.frame} or \code{tibble}. The columns have to be the property
levels, starting with \code{Property}, and the data column(s). See \code{emeScheme_gd} as an example.}

\item{s}{the \code{emeScheme} object to which the data should be added. If the data exists, it will be overwritten, \code{NA} and \code{NULL} values will be treated as not given.}

\item{dataCol}{data column indicator. Can be an index of data column (1 =
first data column, 2 = second, ...), or the name of the data column.
Default = 1.}

\item{verbose}{if \code{TRUE}, progress and diagnostig messages will be printed.}
}
\value{
the emeScheme \code{s} filled in with the values as specified in \code{x}
}
\description{
Title
}
\examples{
addDataToEmeScheme( verbose = TRUE )
}
