% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.dmdScheme_example.R
\docType{data}
\name{dmdScheme_example}
\alias{dmdScheme_example}
\title{dmdScheme_example as imported from the Google sheet.}
\format{An object of class \code{dmdSchemeSet} (inherits from \code{list}) of length 6.}
\usage{
dmdScheme_example
}
\description{
The dataset contains example data. It was created by using the code below.
}
\examples{
\dontrun{
## Created by using
dmdScheme_example <- new_dmdSchemeSet(
   x = dmdScheme_raw,
   keepData = TRUE,
   verbose = TRUE
)
}
}
\keyword{datasets}
