% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmdScheme_example.R
\name{dmdScheme_example}
\alias{dmdScheme_example}
\title{Object of class \code{\link[dmdScheme]{dmdSchemeSet}} containing example data.}
\usage{
dmdScheme_example()
}
\description{
The dataset contains example data. It was created by using the code below.
}
\examples{
dmdScheme_example()

}
