% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmdScheme_raw.R
\name{dmdScheme_raw}
\alias{dmdScheme_raw}
\alias{dmdSchemeSet_raw}
\alias{dmdSchemeData_raw}
\title{Object of class \code{dmdScheme_raw } containing the raw data as read in.}
\usage{
dmdScheme_raw()
}
\description{
The dataset contains raw data. An object of class \code{dmdScheme_raw} is
returned by the function \code{\link[dmdScheme]{read_excel_raw}} with the argument
\code{raw = TRUE} and \code{\link[dmdScheme]{read_excel_raw}}. It is usually an
intermediate object, as in the normal workflow, this object is automatically
converted to an object of class \code{\link[dmdScheme]{dmdSchemeSet}}.
}
\details{
\describe{
  \item{\code{dmdSchemeData_raw}: }{a \code{\link[tibble]{tibble}} as returned by the function \code{\link[readxl]{read_excel}} with the class \code{dmdSchemeData_raw}}
  \item{\code{dmdSchemeSet_raw: }}{a \code{list} with where each element is a \code{dmdSchemeData_raw} object with additional attributes:
    \describe{
      \item{propertyName: }{name of the dmdScheme for which this object contains the raw data. In the spreadsheet, it is in the cell H1 in ther in the Experiment tab (DATA dmeScheme v0.9.5.)}
      \item{dmdSchemeVersion: }{version of the dmdScheme used. In the spreadsheet, it is in the cell H1 in ther in the Experiment tab (DATA dmeScheme v0.9.5.)}
      \item{names: }{the names of the \code{dmdSchemeData_raw} sets. In the spreadsheet, the names of the tabs.}
    }
  }
}
}
\examples{
dmdScheme_raw()

}
