% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmdScheme_versions.R
\name{dmdScheme_versions}
\alias{dmdScheme_versions}
\title{Returns versions of package and dmdScheme}
\usage{
dmdScheme_versions(schemeName = "dmdScheme")
}
\arguments{
\item{schemeName}{name of the scheme. This can be used when more than one
scheme is installed, to query the versions of a specific scheme.}
}
\value{
a named \code{list}, containing the following objects:
\itemize{
  \item{package}{ containing an object of length on of type \code{package_version} with the version of the package}
  \item{dmdScheme}{ containing an object of length on of type \code{numeric_version} with the version of the metadata scheme dmdScheme}
}
}
\description{
The versions of the \code{scheme} and version of the \code{package} are
retrieved from the installed scheme as specified by the argument
\code{schemeName}.
}
\examples{
dmdScheme_versions()
dmdScheme_versions(schemeName = "dmdScheme")
}
