% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmdScheme_versions.R
\name{dmdScheme_versions}
\alias{dmdScheme_versions}
\title{Returns versions of package and dmdScheme}
\usage{
dmdScheme_versions(schemeName = "dmdScheme", lib.loc = NULL)
}
\arguments{
\item{schemeName}{name of the scheme. Default: dmdScheme. Only for developing new schemes needed.}

\item{lib.loc}{a character vector of directory names of R libraries, or NULL.
The default value of NULL corresponds to all libraries currently known. If
the default is used, the loaded packages and namespaces are searched before
the libraries.}
}
\value{
a named \code{list()}, containing the following objects:
\itemize{
  \item{package}{ containing an object of length on of type \code{package_version} with the version of the package}
  \item{dmdScheme}{ containing an object of length on of type \code{numeric_version} with the version of the metadata scheme dmdScheme}
}
}
\description{
Returns versions of package and dmdScheme
}
\examples{
dmdScheme_versions()
}
