% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeScheme_extract.R
\name{emeScheme_extract}
\alias{emeScheme_extract}
\title{Split \code{emeScheme} object into multiple by using \code{DataFileNameMetaData$dataFileName}}
\usage{
emeScheme_extract(dataFile, x)
}
\arguments{
\item{dataFile}{name of dataFileName whose metadata will be extracted from \code{x}. Has to be an exact match, no wildcards are expanded.}

\item{x}{\code{emeScheme} object from which data to extract}
}
\value{
\code{emeScheme} object containing metadata for the data file \code{dataFileName}
}
\description{
One \code{emeScheme} object can contain metadata for multiple datafiles. For
archiving, these should be split into single \code{emeScheme} objects, one
for each \code{DataFileNameMetaData$dataFileName}.
The filtering is done as followed:
\describe{
  \item{DataFileMetaDFata}{\code{DataFileMetaData$dataFileName == dataFile}}
  \item{Treatment}{\code{Treatment$parameter \%in\% DataFileMetaData$mappingColumn} and \code{DataFileMetaData$columnData == "Treatment"} }
  \item{Measurement}{\code{Measurement$name \%in\% DataFileMetaData$mappingColumn} and \code{DataFileMetaData$columnData == "Measurement"} }
}
}
\examples{
emeScheme_extract("smell.csv", emeScheme_example)
## returns the emeScheme data for the data file 'smell.csv'

emeScheme_extract("DoesNotExist", emeScheme_example)
## returns an empty emeScheme
}
