% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeScheme_split.R
\name{emeScheme_split}
\alias{emeScheme_split}
\title{Split \code{emeScheme} object into multiple by using \code{DataFileNameMetaData$dataFileName}}
\usage{
emeScheme_split(x, saveAsType = "none", path = ".")
}
\arguments{
\item{x}{\code{emeScheme} object to be split}

\item{saveAsType}{if given, save the result to files named following the
pattern \code{DATAFILENAME_attr(x, "propertyName").saveAsType}. If missing,
do not save and return the result. Allowed values at the moment:
\itemize{
 \item{none} {do not save the resulting list}
 \item{rds} {save results in files using \code{saveRDS()}}
 \item{xml} {save results in xml files using \code{emeSchemeToXml()}}
 \item{multiple values of the above} {will be saved in all specified formats}
}}

\item{path}{path where the files should be saved to}
}
\value{
if \code{saveAsType} valid and not equal to \code{none}, \code{character} vector
  containing the file names where the splitted metadata has been saved to, if
  \code{saveAsType} missing, \code{list} where each element is one
  \code{emeScheme} object for a data file as specified in
  \code{DataFileNameMetaData$dataFileName}.
}
\description{
One \code{emeScheme} object can contain metadata for multiple datafiles. For
archiving, these should be split into single \code{emeScheme} objects, one
for each \code{DataFileNameMetaData$dataFileName}.
\bold{ATTENTION: Files are overwritten without warning!!!!!}
}
\examples{
emeScheme_split(emeScheme_example)
## x is a list containing all the emeSchemes for each data file

emeScheme_split(emeScheme_example, saveAsType = "rds", path = tempdir())
## saves the resulting object as rds using saveRDS() into the tmpdir()

emeScheme_split(emeScheme_example, saveAsType = "xml", path = tempdir())
## saves the resulting object as xml into the tmpdir()

emeScheme_split(emeScheme_example, saveAsType = c("rds", "xml"), path = tempdir())
## saves the resulting object as rds and xml into the tmpdir()
}
