% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeSchemeToXml.R
\name{emeSchemeToXml}
\alias{emeSchemeToXml}
\title{Convert \code{emeScheme} object to XML}
\usage{
emeSchemeToXml(x, tag, file, output = "metadata", confirmationCode)
}
\arguments{
\item{x}{\code{emeScheme} object to be converted}

\item{tag}{xml tag of the root level. \bold{Has to be "emeScheme for the initial call!}}

\item{file}{empty, \code{NULL} or file name. See details below}

\item{output}{either \code{"metadata"} for export of metadata only or
\code{"complete"} for export including classes et al.}

\item{confirmationCode}{code reported at the end of the validation report
generated by using e.g. \code{validate(..., report = "html")} or any other
reports. No export is possible witholut this \code{confirmationCode}}
}
\value{
dependant on the value of \code{file}. See Details
}
\description{
Converts an \code{emeScheme} object into an xml object.
}
\details{
Depending on the value of \code{file}, the function returns different results:
\itemize{
  \item{\code{file} not given: }{an \code{XMLNode} object}
  \item{\code{file = NULL}: }{a string characterisation of the \code{XMLNode} object}
  \item{\code{file = fileName}: }{saves the string characterisation of the \code{XMLNode} object to an xml file}
}
Based on David LeBauer, Carl Davidson, Rob Kooper. See \url{https://stackoverflow.com/a/27865512/632423}
}
\examples{

## emeSchemeToXml( emeScheme_example, "GoogleData" )
## returns \\code{XMLNode} object

## emeSchemeToXml( emeScheme_example, "GoogleData", file = NULL )
## returns string representation of the \\code{XMLNode} object

}
