% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_new_metadata.R
\name{enter_new_metadata}
\alias{enter_new_metadata}
\title{Enter new metadata to fill a new scheme}
\usage{
enter_new_metadata()
}
\description{
Open \code{emeScheme.xlsx} from system.file("googlesheet",
"emeScheme.xlsx", package = "emeScheme") in excel. New data can be entered
and the file has to be saved at a different location as it is a read-only
file.
}
\examples{
\dontrun{
enter_new_metadata()
}

}
