% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dmdScheme_xlsx.R
\name{format_dmdScheme_xlsx}
\alias{format_dmdScheme_xlsx}
\title{Format the metadata scjheme file}
\usage{
format_dmdScheme_xlsx(fn_org, fn_new, keepData = TRUE)
}
\arguments{
\item{fn_org}{file name of the original excel file to be formated}

\item{fn_new}{file name where the final xlsx should be saved to. If missing, it will not be saved.}

\item{keepData}{if \code{TRUE}, data from data cells will be empty}
}
\value{
invisibly the workbook as a workbook object as created by \code{xlsx.crerateWorkbook()}
}
\description{
Loads  \code{fn_org)}, formats it and saves it as \code{fn_new}.
}
