% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_emeScheme_xlsx.R
\name{format_emeScheme_xlsx}
\alias{format_emeScheme_xlsx}
\title{Format the \code{googlesheet/emeScheme.xlsx}}
\usage{
format_emeScheme_xlsx(emeScheme_gd)
}
\arguments{
\item{emeScheme_gd}{specifying this, makes it possible to use this function even during the update process.}
}
\value{
invisibly \code{NULL}
}
\description{
Takes no arguments. Loads  \code{system.file("inst", "googlesheet", "emeScheme.xlsx", package = "emeScheme")}, formats it and saves it under the same name.
}
