% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_emeScheme_xlsx.R
\name{format_emeScheme_xlsx}
\alias{format_emeScheme_xlsx}
\title{Format the file \code{system.file("emeScheme.xlsx", package = "emeScheme")}}
\usage{
format_emeScheme_xlsx(fn_org, fn_new, keepData = TRUE)
}
\arguments{
\item{fn_org}{file name of the original excel file to be formated}

\item{fn_new}{file name where the final xlsx should be saved to. If missing, it will not be saved.}

\item{keepData}{if \code{TRUE}, data from data cells will be empty}
}
\value{
invisibly the workbook as a workbook object as created by \code{xlsx.crerateWorkbook()}
}
\description{
Takes no arguments. Loads  \code{system.file("emeScheme.xlsx", package =
"emeScheme")}, formats it and saves it under the same name.
}
