% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdToScheme.R
\name{gdToScheme}
\alias{gdToScheme}
\title{Convert the data stored in \code{emeScheme} into a list of tibbles}
\usage{
gdToScheme(x = emeScheme)
}
\value{
\code{list} of \code{list} of ... \code{tibbles}
}
\description{
TODO: Add attriburtes for repeatability and check if tibble is appropriate (hopefully!!!!)
}
\examples{
gdToScheme()

}
