% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example.R
\name{make_example}
\alias{make_example}
\title{Create examples in workingdirectoryt}
\usage{
make_example(name)
}
\arguments{
\item{name}{name of the example}
}
\value{
invisibly \code{NULL}
}
\description{
This function copies the installed example into the working directory so that it can be used.
}
\examples{
make_exapme()
\dontrun{
make_example("basic")
}
}
