% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dmdSchemeSet.R
\name{new_dmdSchemeSet}
\alias{new_dmdSchemeSet}
\title{Convert the data stored in \code{dmdScheme_raw} into a list of tibbles}
\usage{
new_dmdSchemeSet(x, keepData = FALSE, convertTypes = TRUE,
  verbose = FALSE, warnToError = TRUE)
}
\arguments{
\item{x}{object of class \code{dmdSchemeSet_raw} as e.g. returned by \code{read_from_excel(raw = TRUE)}}

\item{keepData}{if the data should be kept or replaced with one row with NAs}

\item{convertTypes}{if \code{TRUE}, the types specified in the types column
are used for the data type. Otherwise, they are left at type \code{character}}

\item{verbose}{give messages to make finding errors in data easier}

\item{warnToError}{if \code{TRUE}, warnings generated during the conversion will raise an error}
}
\value{
\code{list} of \code{list} of ... \code{tibbles}
}
\description{
Convert the data stored in \code{dmdScheme_raw} into a list of tibbles
}
\examples{
new_dmdSchemeSet(dmdScheme_raw, keepData = TRUE)

}
