% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml.R
\name{read_xml}
\alias{read_xml}
\title{Function to read x from an XML file}
\usage{
read_xml(file, ...)
}
\arguments{
\item{file}{Path to file or connection to write to.}

\item{...}{additional parameter for the conversion function \code{as_xml()}}
}
\description{
Read the XML file \code{file} and convert it to a \code{dmdScheme} object using the function \code{as_dmdScheme()}.
}
\examples{
# write_xml(dmdScheme_raw, file = tempfile())

}
