% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml.R
\name{read_xml}
\alias{read_xml}
\title{Function to read x from an XML file}
\usage{
read_xml(file, keepData = TRUE, useSchemeInXml = NULL, verbose = FALSE)
}
\arguments{
\item{file}{Path to file or connection to write to.}

\item{keepData}{if the data should be kept or replaced with one row with NAs}

\item{useSchemeInXml}{if \code{TRUE}, use scheme definition in xml and raise
an error if the xml does not contain a scheme definition. If False, use the
scheme definition from the corresponding installed package, even if the xml
contains a scheme definition. if \code{NULL} (the default), use the
definition in the xml if it contains a definition, if not use the
corresponding definition from the installed package.}

\item{verbose}{give verbose progress info. Useful for debugging.}
}
\description{
Read the XML file \code{file} and convert it to a \code{dmdScheme} object using the function \code{as_dmdScheme()}.
}
\examples{
# write_xml(dmdScheme_raw(), file = tempfile())

}
