% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme_list.R, R/scheme_use.R
\name{scheme_list}
\alias{scheme_list}
\alias{scheme_use}
\title{List installed schemes for this package}
\usage{
scheme_list()

scheme_use(schemeDefinition = "dmdScheme.xml")
}
\arguments{
\item{schemeDefinition}{}
}
\value{
list of schemes installed in the folder \code{system.file("schemeDefinitions", package = "dmdScheme")}
}
\description{
Schemes can be added by using the \link{scheme_add()} function.

Switch from the current scheme to a new scheme as defined in the file
\code{schemeDefinition}. The name of the scheme definition as in the folder \code{}.
New schemes can be added to the library via a call to \code{scheme_add()}.
}
\examples{
scheme_list()
use_scheme("dmdScheme.xml)
use_scheme("dmdScheme.xlsx)
\dontrun{
scheme_add("path/to/definition.xml)
scheme_add("path/to/definition.xlsx)
}
}
