% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme_make.R
\name{scheme_make}
\alias{scheme_make}
\title{Functions to manage schemes}
\usage{
scheme_make(schemeDefinition, examples = NULL,
  install_R_package = NULL, path = ".", overwrite = FALSE)
}
\arguments{
\item{schemeDefinition}{path to the \code{.xlsx} file containing the
\bold{definition} of the scheme as well as the \bold{example}}

\item{examples}{character vector of directories which should be included as
examples. The name of the director will be the name of the example. The
example can contain a file with the name \code{EXAMPLENAME.html} where
\code{EXAMPLENAME} is the name of the folder. This html will abe
automatically opened when calling \code{make_example("EXAMPLENAME")}
Otherwise there are no restrictions to formats.}

\item{install_R_package}{path to the R script to install the R package. If \code{NULL},
no command is given.}

\item{path}{where the final scheme definition should be created.}

\item{overwrite}{if \code{TRUE}, the scheme definition in \code{path} will be
overwritten.}
}
\value{
fully qualified path to the created scheme
}
\description{
Functions to manage schemes
}
