% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFromGoogleSheet.R
\name{updateFromGoogleSheet}
\alias{updateFromGoogleSheet}
\title{Update data from googlesheets}
\usage{
updateFromGoogleSheet(token = NULL, force = FALSE)
}
\arguments{
\item{token}{token for google access. If \code{NULL} (default), user will be asked}

\item{force}{force update even if no newer version on google docs}
}
\value{
a googlesheet object
}
\description{
Update the data and increase the version in the DESCRIPTION if it has changed.
}
