% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFromNewSheet.R
\name{updateFromNewSheet}
\alias{updateFromNewSheet}
\title{Update data from emeScheme.xlsx --- ONLY FOR DEVLOPMENT NEEDED}
\usage{
updateFromNewSheet(force = FALSE)
}
\arguments{
\item{force}{force update even if \code{here::here("inst", "emeScheme.xlsx")}
has the same md5 suma s in the DESCRIPTION}
}
\value{
invisibly NULL
}
\description{
Update the data from the file \code{here::here("inst", "emeScheme.xlsx")}
and bump the version in the DESCRIPTION if it has changed.
}
