% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_error_functions.R
\name{valErr_extract}
\alias{valErr_extract}
\title{Extract all fields named error of class \code{dmdScheme_validation}}
\usage{
valErr_extract(x, returnRootError = FALSE)
}
\arguments{
\item{x}{object of class \code{dmdScheme_validation}}

\item{returnRootError}{if \code{TRUE}, return all errors \bold{including} the error in the object x.}
}
\value{
named numeric vector of the error levels of the different validations done
}
\description{
Extract all fields named error of class \code{dmdScheme_validation}
}
