% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate structure of object of class \code{emeSchemeSet_raw} against \code{emeScheme}.}
\usage{
validate(x, path = "", validateData = TRUE, report = "none",
  report_author = "Tester",
  report_title = "Validation of data against emeScheme",
  errorIfFalse = TRUE)
}
\arguments{
\item{x}{object of class \code{emeSchemeSet_raw} as returned from \code{read_from_excel( keepData = FALSE, raw = TRUE)}}

\item{path}{path to the data files}

\item{validateData}{if \code{TRUE} data is validated as well; the structure is always validated}

\item{report}{determines if and in which format a report of the validation should be generated. Allowed values are:
\itemize{
  \item{\bold{missing}  : } {no report is generated (the \code{default})}
  \item{\bold{'html'}   : } {a html (.html) report is generated and opened}
  \item{\bold{'pdf'}    : } {a pdf (.pdf) report is generated and opened}
  \item{\bold{'word'}   : } {a word (.docx)report is generated and opened}
  \item{\bold{'all'}    : } {all of the defined reports}
}}

\item{report_author}{name of the author for the report}

\item{report_title}{title of the report for the report}

\item{errorIfFalse}{if \code{TRUE} an error will be raised if the schemes are not identical, i.e. there are structural differences.}
}
\value{
return invisibly the report in form of a \code{list}
}
\description{
Validate structure of object of class \code{emeSchemeSet_raw} against \code{emeScheme}.
}
\examples{
validate( emeScheme_raw )

}
