% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xml.R
\name{write_xml}
\alias{write_xml}
\title{Write write x as an XML file to disk}
\usage{
write_xml(x, file, output = "metadata", ...)
}
\arguments{
\item{x}{object which will be converted to and saved as an xml file.}

\item{file}{Path to file or connection to write to.}

\item{output}{specifies the content and format of the exported xml. see \link{as_xml} for details}

\item{...}{additional parameter for the conversion function \code{as_xml}}
}
\description{
Convert the object \code{x} to an \code{xml_document} object using the function \code{as_xml()} and write it to a file.
If no method \code{as_xml()} exists for the object |class{x}, an error will be raised.
}
\examples{
write_xml(dmdScheme(), file = tempfile())

}
