% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_dmdScheme.R
\name{xml_to_dmdScheme}
\alias{xml_to_dmdScheme}
\title{Convert xml to \code{dmdScheme}}
\usage{
xml_to_dmdScheme(xml, verbose = FALSE)
}
\arguments{
\item{xml}{either xml object or text represenatation from text object}

\item{verbose}{default: \code{FALSE}; give messages to make finding errors in
data easier}
}
\value{
\code{dmdScheme} or descendant object
}
\description{
TRhe scheme is automatically chosen from the propety \code{propertyName} in
the xml file. An error is raised if it does not exist.
}
\details{
The resulting \code{dmdScheme} does only contain the non-missing values which
are specified in the xml file.
}
\examples{
## x <- dmdScheme_example

}
