% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_dmdScheme.R
\name{xml_to_dmdScheme}
\alias{xml_to_dmdScheme}
\title{Convert a exported xml file to a \code{dmdScheme_Set} object}
\usage{
xml_to_dmdScheme(file, onlyScheme = FALSE, verbose = FALSE)
}
\arguments{
\item{file}{a file containing the xml or an xml object}

\item{verbose}{give verbose progress info. Useful for debugging.}
}
\value{
\code{dmdScheme} or descendant object
}
\description{
The scheme is automatically chosen from the propety \code{propertyName} in
the xml file. Nevertheless, the scheme has to be installed, but it does not
hav to be loaded. An error is raised if it does not exist.
}
\examples{
## x <- dmdScheme_example

}
