% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlToEmeScheme.R
\name{xmlToEmeScheme}
\alias{xmlToEmeScheme}
\title{Convert xml to \code{emeScheme}}
\usage{
xmlToEmeScheme(xml, verbose = FALSE)
}
\arguments{
\item{xml}{either xml object or text represenatation from text object}

\item{verbose}{default: \code{FALSE}; give messages to make finding errors in data easier}
}
\value{
\code{emeScheme} object
}
\description{
The resulting \code{emeScheme} does only contain the non-missing values which
are specified in the xml file.
}
\examples{
x <- emeScheme_example

}
